+++ cmd/neofs-ir/defaults_test.go
@@ -26,14 +26,23 @@ func TestValidateDefaultConfig(t *testing.T) {
 				ReconnectionsNumber: 5,
 				ReconnectionsDelay:  5000000000,
 			},
-			Validators: keys.PublicKeys{},
+			DisableAutodeploy: false,
+			Validators:        keys.PublicKeys{},
 		},
-		FSChainAutodeploy: false,
-		NNS:               config.NNS{SystemEmail: ""},
-		Mainnet: config.BasicChain{
-			DialTimeout:         60000000000,
-			ReconnectionsNumber: 5,
-			ReconnectionsDelay:  5000000000,
+		NNS: config.NNS{SystemEmail: ""},
+		Mainnet: config.Mainnet{
+			Enabled:               false,
+			DisableGovernanceSync: false,
+			ExtraFee:              50000000,
+			Contracts: config.Contracts{
+				NeoFS:      "",
+				Processing: "",
+			},
+			BasicChain: config.BasicChain{
+				DialTimeout:         60000000000,
+				ReconnectionsNumber: 5,
+				ReconnectionsDelay:  5000000000,
+			},
 		},
 		Control: config.Control{
 			AuthorizedKeys: keys.PublicKeys{},
@@ -45,7 +54,6 @@ func TestValidateDefaultConfig(t *testing.T) {
 				Path: ".neofs-ir-state",
 			},
 		},
-		Fee: config.Fee{MainChain: 50000000},
 		Timers: config.Timers{
 			CollectBasicIncome: config.BasicTimer{Mul: 1, Div: 2},
 		},