+++ pkg/innerring/config/config.go
@@ -22,7 +22,7 @@ type Config struct {
 
 	NNS NNS `mapstructure:"nns"`
 
-	Mainnet BasicChain `mapstructure:"mainnet"`
+	Mainnet Mainnet `mapstructure:"mainnet"`
 
 	Control Control `mapstructure:"control"`
 
@@ -129,6 +129,15 @@ type Settlement struct {
 // Experimental configures experimental features.
 type Experimental struct {
 	ChainMetaData bool `mapstructure:"chain_meta_data"`
+}
+
+// Mainnet configures mainnet chain settings.
+type Mainnet struct {
+	Enabled               bool      `mapstructure:"enabled"`
+	DisableGovernanceSync bool      `mapstructure:"disable_governance_sync"`
+	ExtraFee              int64     `mapstructure:"extra_fee"`
+	Contracts             Contracts `mapstructure:"contracts"`
+	BasicChain            `mapstructure:",squash"`
 }
 
 // IsSet checks if the key is set in the config.