+++ CHANGELOG.md
@@ -8,12 +8,26 @@ Changelog for NeoFS Node
 ### Fixed
 
 ### Changed
+- Move `fschain_autodeploy` into `fschain.disable_autodeploy` in IR config (#3619)
+- Move `without_mainnet` into `mainnet.enabled` in IR config (#3619)
+- Move `governance.disable` into `mainnet.disable_governance_sync` in IR config (#3619)
+- Move `fee.main_chain` into `mainnet.extra_fee` in IR config (#3619)
+- Move `contracts` into `mainnet.contracts` in IR config (#3619)
 
 ### Removed
 
 ### Updated
 
 ### Updating from v0.49.0
+Use IR configuration options:
+- `fschain.disable_autodeploy` instead of deprecated `fschain_autodeploy` with reverted value,
+by default autodeploy is enabled now;
+- `mainnet.enabled` instead of deprecated `without_mainnet` with reverted value,
+by default mainnet is disabled now;
+- `mainnet.disable_governance_sync` instead of deprecated `governance.disable`;
+- `mainnet.extra_fee` instead of deprecated `fee.main_chain`;
+- `mainnet.contracts` instead of deprecated `contracts`;
+Old options are still supported but will be removed in future releases.
 
 ## [0.49.0] - 2025-10-06 - Dochodo